requirements.txt
diff-match-patch==20200713
Django==5.0.4
django-crispy-forms==1.14.0
django-markdownify==0.9.2
django-role-permissions==3.1.1
django-simple-history==3.2.0
django-taggit==5.0.1
ffmpeg-python==0.2.0
markdown2==2.4.7
olefile==0.46
Pillow==9.4.0
pytz==2022.7.1
requests==2.31.0
six==1.16.0
django-ninja-jwt==5.3.0
orjson==3.9.9
django-cors-headers==4.3.0